move crate to code_model_api
